requirements.txt
azure-functions==1.17.0
python-telegram-bot==20.7 
python 3.12.0
